+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Current Develop Branch
 
+## 7.2.3 Fri Oct 04 2019
+- [#7252](https://github.com/MetaMask/metamask-extension/pull/7252): Fix gas limit when sending tx without data to a contract
+- [#7260](https://github.com/MetaMask/metamask-extension/pull/7260): Do not transate on seed phrases
+
 ## 7.2.2 Tue Sep 24 2019
 - [#7213](https://github.com/MetaMask/metamask-extension/pull/7213): Update minimum Firefox verison to 56.0